[ new ] PHOAS-like combinators for the IL
This commit introduces a PHOAS-inspired set of combinators to build
terms of the internal syntax. This way we do not have to explicitly
write down the `Fin` corresponding to a given bound variable if we
do not want to.

I have added a couple of examples in Test.hs but the gist of it is
that we can use Haskell binders to represent binders in our object
language, the typeclass resolution machinery will figure out how
much shifting needs to happen. E.g. the S, K, I combinators become:

S = lam $ \ f -> lam $ \ g -> lam $ \ x ->
    En (App (App (var f) (var' x)) (En (App (var g) (var' x))))

K = lam $ \ x -> Lam $ var' x

I = lam var'