[ new ] PHOAS for the internal language
This commit introduces a PHOAS for terms of the internal syntax. This
way we do not have to explicitly write down the `Fin` corresponding
to a given bound variable if we do not want to.

I have added a couple of examples in Test.hs but the gist of it is
that we can use Haskell binders to represent binders in our object
language. E.g. the S, K, I combinators become:

S = lam $ \ f -> lam $ \ g -> lam $ \ x ->
    En (App (App (var f) (var' x)) (En (App (var g) (var' x))))

K = lam $ \ x -> Lam $ var' x

I = lam var'